+++ src/main/java/com/xstock/plutus/v1/news/News.java
@@ -28,9 +28,6 @@ public class News {
     @Column(name = "price_change_ratio")
     private Float priceChangeRatio;
 
-//    @Column(name = "weekly_price_change_ratio")
-//    private Float weeklyPriceChangeRatio;
-
     @Column(name = "monthly_price_change_ratio")
     private Float MonthlyPriceChangeRatio;
 
@@ -43,7 +40,7 @@ public class News {
     @Column(name = "publish_date")
     private OffsetDateTime publishDate;
 
-    private Double rsi;
+    private Float rsi;
 
-    private Double rs;
+    private Float rs;
 }
